scripts: setup_cameras: Fix OV5640 format to UYVY 1x16
For CSI-2, the correct formats to use for YUV422 are of the form 1x16
as it is a serial protocol. The formats of the form 2x8 are used for
parallel protocols like DVP and 2x8 formats are deprecated in the
CSI2RX bridge driver thus fix OV5640 format accordingly to use 1X16
instead.

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>